requirements.txt
asgiref==3.11.0
django==5.2.8
django-cors-headers>=4.4.0,<5.0
djangorestframework>=3.15.2,<4.0
djangorestframework-simplejwt>=5.3.1,<6.0
Pillow>=10.4.0,<11.0
sqlparse==0.5.3
tzdata==2025.2
